feature: update credit aggrements on dropdown open
