+++ merchant-site/product-page.html
@@ -174,7 +174,7 @@
       </div>
     </div>
   </div>
-  <script src="../widget/build/static/js/main.c73eec68.js"></script>
+  <script src="../widget/build/static/js/main.0c34b510.js"></script>
   <script src="../widget/build/static/js/453.234e652c.chunk.js"></script>
 
 </body>
